Support sensor position substrings
For, e.g, unilaterally worn sensors, select the matching sensor position. For the wrist sensor,
the position will for example be 'LeftWrist' for most participants, 'RightWrist' for some others.
When creating a dataset, a more flexible way (e.g., specifying 'Wrist') to define the sensor position
is required in those cases. This code of this commit searches for the sensor position that contains
the sensor_pos string.